requirements.txt
google-generativeai
python-dotenv
flask==3.0.3
PyMuPDF
